+++ flotils/convenience.py
@@ -6,7 +6,7 @@ from __future__ import unicode_literals
 
 __author__ = "d01"
 __email__ = "jungflor@gmail.com"
-__copyright__ = "Copyright (C) 2018, Florian JUNG"
+__copyright__ = "Copyright (C) 2018-19, Florian JUNG"
 __license__ = "MIT"
 __version__ = "0.1.0"
 __date__ = "2018-01-27"