+++ flotils/loadable.py
@@ -9,18 +9,17 @@ Module for loading/saving data/classes with json
 
 __author__ = "the01"
 __email__ = "jungflor@gmail.com"
-__copyright__ = "Copyright (C) 2013-18, Florian JUNG"
+__copyright__ = "Copyright (C) 2013-19, Florian JUNG"
 __license__ = "MIT"
-__version__ = "0.3.1"
-__date__ = "2018-01-27"
+__version__ = "0.4.0"
+__date__ = "2019-03-26"
 # Created: 2014-08-29 09:38
 
 import os
 import datetime
 import json
+import io
 
-import dateutil.parser
-from dateutil.tz import tzutc
 import yaml
 
 from .logable import Logable, ModuleLogable
@@ -31,7 +30,6 @@ class Logger(ModuleLogable):
 
 
 logger = Logger()
-UTC = tzutc()
 
 
 # TODO: New format according to https://gist.github.com/majgis/4200488
@@ -42,7 +40,7 @@ class DateTimeEncoder(json.JSONEncoder):
         if isinstance(obj, datetime.datetime):
             # save all without tz info as UTC
             if obj.tzinfo:
-                obj = obj.astimezone(UTC).replace(tzinfo=None)
+                obj = (obj - obj.tzinfo.utcoffset(obj)).replace(tzinfo=None)
             return {'__datetime__': obj.isoformat() + "Z"}
         elif isinstance(obj, datetime.date):
             return {'__date__': obj.isoformat()}
@@ -66,27 +64,39 @@ class DateTimeDecoder(object):
 
     @staticmethod
     def _as_datetime(dct):
-        if u"__datetime__" in dct.keys():
+        if "__datetime__" in dct.keys():
             # Should be UTC
-            return dateutil.parser.parse(
-                dct['__datetime__']
-            ).replace(tzinfo=UTC)
+            return datetime.datetime.strptime(
+                dct['__datetime__'], "%Y-%m-%dT%H:%M:%S.%fZ"
+            )
         raise TypeError("Not Datetime")
 
     @staticmethod
     def _as_date(dct):
         if "__date__" in dct:
-            return dateutil.parser.parse(dct['__date__']).date()
+            d = datetime.datetime.strptime(
+                dct['__date__'], "%Y-%m-%d"
+            )
+            if d:
+                return d.date()
+            return d
         raise TypeError("Not Date")
 
     @staticmethod
     def _as_time(dct):
         if "__time__" in dct:
-            return dateutil.parser.parse(dct['__time__']).time()
+            d = datetime.datetime.strptime(
+                dct['__time__'], "%H:%M:%S.%f"
+            )
+            if d:
+                return d.time()
+            return d
         raise TypeError("Not Time")
 
     @staticmethod
     def decode(dct):
+        if not isinstance(dct, dict):
+            return dict
         if "__type__" in dct:
             obj_type = dct.pop('__type__')
 
@@ -107,101 +117,6 @@ class DateTimeDecoder(object):
                     return dct
 
 
-def loadJSON(json_data, decoder=None):
-    """
-    Load data from json string
-
-    :param json_data: Stringified json object
-    :type json_data: str | unicode
-    :param decoder: Use custom json decoder
-    :type decoder: T <= DateTimeDecoder
-    :return: Json data
-    :rtype: None | int | float | str | list | dict
-    """
-    # deprecated
-    import warnings
-    warnings.warn(
-        "This method is no longer in use - Please use load_json() instead",
-        DeprecationWarning
-    )
-    return load_json(json_data, decoder)
-
-
-def loadJSONFile(json_data, decoder=None):
-    """
-    Load data from json file
-
-    :param json_data: Readable object or path to file
-    :type json_data: FileIO | str
-    :param decoder: Use custom json decoder
-    :type decoder: T <= DateTimeDecoder
-    :return: Json data
-    :rtype: None | int | float | str | list | dict
-    """
-    # deprecated
-    import warnings
-    warnings.warn(
-        "This method is no longer in use - Please use load_json_file() instead",
-        DeprecationWarning
-    )
-    return load_json_file(json_data, decoder)
-
-
-def saveJSON(val, pretty=False, sort=True, encoder=None):
-    """
-    Save data to json string
-
-    :param val: Value or struct to save
-    :type val: None | int | float | str | list | dict
-    :param pretty: Format data to be readable (default: False)
-                    otherwise going to be compact
-    :type pretty: bool
-    :param sort: Sort keys (default: True)
-    :type sort: bool
-    :param encoder: Use custom json encoder
-    :type encoder: T <= DateTimeEncoder
-    :return: The jsonified string
-    :rtype: str | unicode
-    """
-    # deprecated
-    import warnings
-    warnings.warn(
-        "This method is no longer in use - Please use save_json() instead",
-        DeprecationWarning
-    )
-    return save_json(val, pretty, sort, encoder)
-
-
-def saveJSONFile(
-        json_data, val,
-        pretty=False, compact=True, sort=True, encoder=None
-):
-    """
-    Save data to json file
-
-    :param json_data: Writable object or path to file
-    :type json_data: FileIO | str
-    :param val: Value or struct to save
-    :type val: None | int | float | str | list | dict
-    :param pretty: Format data to be readable (default: False)
-    :type pretty: bool
-    :param compact: Format data to be compact (default: True)
-    :type compact: bool
-    :param sort: Sort keys (default: True)
-    :type sort: bool
-    :param encoder: Use custom json encoder
-    :type encoder: T <= DateTimeEncoder
-    :rtype: None
-    """
-    # deprecated
-    import warnings
-    warnings.warn(
-        "This method is no longer in use - Please use save_json_file() instead",
-        DeprecationWarning
-    )
-    return save_json_file(json_data, val, pretty, compact, sort, encoder)
-
-
 def load_json(json_data, decoder=None):
     """
     Load data from json string
@@ -218,12 +133,12 @@ def load_json(json_data, decoder=None):
     return json.loads(json_data, object_hook=decoder.decode)
 
 
-def load_json_file(json_data, decoder=None):
+def load_json_file(file, decoder=None):
     """
     Load data from json file
 
-    :param json_data: Readable object or path to file
-    :type json_data: FileIO | str
+    :param file: Readable object or path to file
+    :type file: FileIO | str
     :param decoder: Use custom json decoder
     :type decoder: T <= DateTimeDecoder
     :return: Json data
@@ -231,10 +146,10 @@ def load_json_file(json_data, decoder=None):
     """
     if decoder is None:
         decoder = DateTimeDecoder
-    if not hasattr(json_data, "read"):
-        with open(json_data, "rb") as f:
+    if not hasattr(file, "read"):
+        with io.open(file, "r", encoding="utf-8") as f:
             return json.load(f, object_hook=decoder.decode)
-    return json.load(json_data, object_hook=decoder.decode)
+    return json.load(file, object_hook=decoder.decode)
 
 
 def save_json(val, pretty=False, sort=True, encoder=None):
@@ -272,14 +187,14 @@ def save_json(val, pretty=False, sort=True, encoder=None):
 
 
 def save_json_file(
-        json_data, val,
+        file, val,
         pretty=False, compact=True, sort=True, encoder=None
 ):
     """
     Save data to json file
 
-    :param json_data: Writable object or path to file
-    :type json_data: FileIO | str
+    :param file: Writable object or path to file
+    :type file: FileIO | str | unicode
     :param val: Value or struct to save
     :type val: None | int | float | str | list | dict
     :param pretty: Format data to be readable (default: False)
@@ -297,15 +212,15 @@ def save_json_file(
         encoder = DateTimeEncoder
     opened = False
 
-    if not hasattr(json_data, "write"):
-        json_data = open(json_data, "wb")
+    if not hasattr(file, "write"):
+        file = io.open(file, "w", encoding="utf-8")
         opened = True
 
     try:
         if pretty:
             json.dump(
                 val,
-                json_data,
+                file,
                 indent=4,
                 separators=(',', ': '),
                 sort_keys=sort,
@@ -314,16 +229,17 @@ def save_json_file(
         elif compact:
             json.dump(
                 val,
-                json_data,
+                file,
                 separators=(',', ':'),
                 sort_keys=sort,
                 cls=encoder
             )
         else:
-            json.dump(val, json_data, sort_keys=sort, cls=encoder)
+            json.dump(val, file, sort_keys=sort, cls=encoder)
     finally:
         if opened:
-            json_data.close()
+            file.close()
+
 
 def load_yaml(data):
     """
@@ -337,19 +253,19 @@ def load_yaml(data):
     return yaml.load(data)
 
 
-def load_yaml_file(data):
+def load_yaml_file(file):
     """
     Load data from yaml file
 
-    :param json_data: Readable object or path to file
-    :type json_data: FileIO | str | unicode
+    :param file: Readable object or path to file
+    :type file: FileIO | str | unicode
     :return: Yaml data
     :rtype: None | int | float | str | unicode | list | dict
     """
-    if not hasattr(data, "read"):
-        with open(data, "rb") as f:
+    if not hasattr(file, "read"):
+        with io.open(file, "r", encoding="utf-8") as f:
             return yaml.load(f)
-    return yaml.load(data)
+    return yaml.load(file)
 
 
 def save_yaml(val):
@@ -364,26 +280,26 @@ def save_yaml(val):
     return yaml.dump(val)
 
 
-def save_yaml_file(yaml_data, val):
+def save_yaml_file(file, val):
     """
     Save data to yaml file
 
-    :param yaml_data: Writable object or path to file
-    :type yaml_data: FileIO | str | unicode
+    :param file: Writable object or path to file
+    :type file: FileIO | str | unicode
     :param val: Value or struct to save
     :type val: None | int | float | str | unicode | list | dict
     """
     opened = False
 
-    if not hasattr(yaml_data, "write"):
-        yaml_data = open(yaml_data, "wb")
+    if not hasattr(file, "write"):
+        file = io.open(file, "w", encoding="utf-8")
         opened = True
 
     try:
-        yaml.dump(val, yaml_data)
+        yaml.dump(val, file)
     finally:
         if opened:
-            yaml_data.close()
+            file.close()
 
 
 def load_file(path):
@@ -405,7 +321,7 @@ def load_file(path):
         raise IOError("File not found {}".format(path))
 
     try:
-        with open(path, "rb") as f:
+        with io.open(path, "r", encoding="utf-8") as f:
             if path.endswith(".json"):
                 res = load_json_file(f)
             elif path.endswith(".yaml") or path.endswith(".yml"):
@@ -434,7 +350,7 @@ def save_file(path, data, readable=False):
         IOError("No path specified to save")
 
     try:
-        with open(path, "wb") as f:
+        with io.open(path, "w", encoding="utf-8") as f:
             if path.endswith(".json"):
                 save_json_file(
                     f,
@@ -451,37 +367,17 @@ def save_file(path, data, readable=False):
         raise IOError(e)
 
 
-def joinPathPrefix(path, pre_path=None):
-    """
-        If path set and not absolute, append it to pre path (if used)
-
-        :param path: path to append
-        :type path: str | None
-        :param pre_path: Base path to append to (default: None)
-        :type pre_path: None | str
-        :return: Path or appended path
-        :rtype: str | None
-        """
-    # deprecated
-    import warnings
-    warnings.warn(
-        "This method is no longer in use - Please use join_path_prefix() instead",
-        DeprecationWarning
-    )
-    return join_path_prefix(path, pre_path)
-
-
 def join_path_prefix(path, pre_path=None):
     """
-        If path set and not absolute, append it to pre path (if used)
+    If path set and not absolute, append it to pre path (if used)
 
-        :param path: path to append
-        :type path: str | None
-        :param pre_path: Base path to append to (default: None)
-        :type pre_path: None | str
-        :return: Path or appended path
-        :rtype: str | None
-        """
+    :param path: path to append
+    :type path: str | None
+    :param pre_path: Base path to append to (default: None)
+    :type pre_path: None | str
+    :return: Path or appended path
+    :rtype: str | None
+    """
     if not path:
         return path
 
@@ -509,16 +405,16 @@ class Loadable(Logable):
             settings = {}
         super(Loadable, self).__init__(settings)
         sett_path = settings.get('settings_file', None)
-        self._prePath = settings.get('path_prefix', None)
+        self._pre_path = settings.get('path_prefix', None)
 
         if sett_path:
             sett_path = self.join_path_prefix(sett_path)
             sett = self.load_settings(sett_path)
             sett_prepath = sett.get('path_prefix')
             if sett_prepath:
-                # if settPath is absolute path
-                # -> set to settPath else join with dict path_prefix
-                self._prePath = self.join_path_prefix(sett_prepath)
+                # if sett_path is absolute path
+                # -> set to sett_path else join with dict path_prefix
+                self._pre_path = self.join_path_prefix(sett_prepath)
             # settings in constructor overwrite settings from file
             sett.update(settings)
             settings.update(sett)
@@ -526,84 +422,68 @@ class Loadable(Logable):
             # to apply loaded settings to logable as well
             super(Loadable, self).__init__(settings)
 
-    def joinPathPrefix(self, path):
-        """
-        If path set and not absolute, append it to self._prePath
-
-        :param path: Path to append
-        :type path: str | None
-        :return: Path or appended path
-        :rtype: str | None
-        """
+    @property
+    def _prePath(self):
         import warnings
         warnings.warn(
-            "This method is no longer in use - Please use .join_path_prefix() instead",
+            "This variable is no longer in use - Please use _pre_path instead",
             DeprecationWarning
         )
-        return self.join_path_prefix(path)
+        return self._pre_path
+
+    @_prePath.setter
+    def set_prePath(self, value):
+        import warnings
+        warnings.warn(
+            "This variable is no longer in use - Please use _pre_path instead",
+            DeprecationWarning
+        )
+        self._pre_path = value
 
     def join_path_prefix(self, path):
         """
-        If path set and not absolute, append it to self._prePath
+        If path set and not absolute, append it to self._pre_path
 
         :param path: Path to append
         :type path: str | None
         :return: Path or appended path
         :rtype: str | None
         """
-        return join_path_prefix(path, self._prePath)
+        return join_path_prefix(path, self._pre_path)
 
-    def _loadJSONFile(self, json_data, decoder=None):
+    def _load_json_file(self, file, decoder=None):
         """
         Load data from json file
 
-        :param json_data: Readable file or path to file
-        :type json_data: FileIO | str
+        :param file: Readable file or path to file
+        :type file: FileIO | str | unicode
         :param decoder: Use custom json decoder
-        :type decoder: T <= DateTimeDecoder
-        :return: Json data
-        :rtype: None | int | float | str | list | dict
-        :raises IOError: Failed to load
-        """
-        import warnings
-        warnings.warn(
-            "This method is no longer in use - Please use ._load_json_file() instead",
-            DeprecationWarning
-        )
-        return self._load_json_file(json_data, decoder)
-
-    def _load_json_file(self, json_data, decoder=None):
-        """
-        Load data from json file
-
-        :param json_data: Readable file or path to file
-        :type json_data: FileIO | str
-        :param decoder: Use custom json decoder
-        :type decoder: T <= DateTimeDecoder
+        :type decoder: T <= flotils.loadable.DateTimeDecoder
         :return: Json data
         :rtype: None | int | float | str | list | dict
         :raises IOError: Failed to load
         """
         try:
-            res = load_json_file(json_data, decoder=decoder)
+            res = load_json_file(file, decoder=decoder)
         except ValueError as e:
             if "{}".format(e) == "No JSON object could be decoded":
                 raise IOError("Decoding JSON failed")
-            self.exception(u"Failed to load from {}".format(json_data))
+            self.exception("Failed to load from {}".format(file))
             raise IOError("Loading file failed")
         except:
-            self.exception(u"Failed to load from {}".format(json_data))
+            self.exception("Failed to load from {}".format(file))
             raise IOError("Loading file failed")
         return res
 
-    def _saveJSONFile(self, json_data, val,
-                      pretty=False, compact=True, sort=True, encoder=None
+    def _save_json_file(
+        self, file, val,
+        pretty=False, compact=True, sort=True, encoder=None
     ):
         """
         Save data to json file
 
-        :param json_data: Writable file or path to file
-        :type json_data: FileIO | str
+        :param file: Writable file or path to file
+        :type file: FileIO | str | unicode
         :param val: Value or struct to save
         :type val: None | int | float | str | list | dict
         :param pretty: Format data to be readable (default: False)
@@ -613,94 +493,48 @@ class Loadable(Logable):
         :param sort: Sort keys (default: True)
         :type sort: bool
         :param encoder: Use custom json encoder
-        :type encoder: T <= DateTimeEncoder
-        :rtype: None
-        :raises IOError: Failed to save
-        """
-        import warnings
-        warnings.warn(
-            "This method is no longer in use - Please use ._save_json_file() instead",
-            DeprecationWarning
-        )
-        return self._save_json_file(json_data, val, pretty, compact, sort, encoder)
-
-    def _save_json_file(self, json_data, val,
-                      pretty=False, compact=True, sort=True, encoder=None
-    ):
-        """
-        Save data to json file
-
-        :param json_data: Writable file or path to file
-        :type json_data: FileIO | str
-        :param val: Value or struct to save
-        :type val: None | int | float | str | list | dict
-        :param pretty: Format data to be readable (default: False)
-        :type pretty: bool
-        :param compact: Format data to be compact (default: True)
-        :type compact: bool
-        :param sort: Sort keys (default: True)
-        :type sort: bool
-        :param encoder: Use custom json encoder
-        :type encoder: T <= DateTimeEncoder
+        :type encoder: T <= flotils.loadable.DateTimeEncoder
         :rtype: None
         :raises IOError: Failed to save
         """
         try:
-            save_json_file(json_data, val, pretty, compact, sort, encoder)
+            save_json_file(file, val, pretty, compact, sort, encoder)
         except:
-            self.exception(u"Failed to save to {}".format(json_data))
+            self.exception("Failed to save to {}".format(file))
             raise IOError("Saving file failed")
 
-    def _load_yaml_file(self, data):
+    def _load_yaml_file(self, file):
         """
         Load data from yaml file
 
-        :param data: Readable object or path to file
-        :type data: FileIO | str | unicode
+        :param file: Readable object or path to file
+        :type file: FileIO | str | unicode
         :return: Yaml data
         :rtype: None | int | float | str | unicode | list | dict
         :raises IOError: Failed to load
         """
         try:
-            res = load_yaml_file(data)
+            res = load_yaml_file(file)
         except:
-            self.exception(u"Failed to load from {}".format(data))
+            self.exception("Failed to load from {}".format(file))
             raise IOError("Loading file failed")
         return res
 
-    def _save_yaml_file(self, yaml_data, val):
+    def _save_yaml_file(self, file, val):
         """
         Save data to yaml file
 
-        :param yaml_data: Writable object or path to file
-        :type yaml_data: FileIO | str | unicode
+        :param file: Writable object or path to file
+        :type file: FileIO | str | unicode
         :param val: Value or struct to save
         :type val: None | int | float | str | unicode | list | dict
         :raises IOError: Failed to save
         """
         try:
-            save_yaml_file(yaml_data, val)
+            save_yaml_file(file, val)
         except:
-            self.exception(u"Failed to save to {}".format(yaml_data))
+            self.exception("Failed to save to {}".format(file))
             raise IOError("Saving file failed")
-
-    def loadSettings(self, path):
-        """
-        Load settings dict
-
-        :param path: Path to settings file
-        :type path: str | unicode
-        :return: Loaded settings
-        :rtype: dict
-        :raises IOError: If file not found or error accessing file
-        :raises TypeError: Settings file does not contain dict
-        """
-        import warnings
-        warnings.warn(
-            "This method is no longer in use - Please use .load_settings() instead",
-            DeprecationWarning
-        )
-        return self.load_settings(path)
 
     def load_settings(self, path):
         """
@@ -717,27 +551,6 @@ class Loadable(Logable):
         if not isinstance(res, dict):
             raise TypeError("Expected settings to be dict")
         return res
-
-    def saveSettings(self, path, settings, readable=False):
-        """
-        Save settings to file
-
-        :param path: File path to save
-        :type path: str | unicode
-        :param settings: Settings to save
-        :type settings: dict
-        :param readable: Format file to be human readable (default: False)
-        :type readable: bool
-        :rtype: None
-        :raises IOError: If empty path or error writing file
-        :raises TypeError: Settings is not a dict
-        """
-        import warnings
-        warnings.warn(
-            "This method is no longer in use - Please use .save_settings() instead",
-            DeprecationWarning
-        )
-        return self.save_settings(path, settings, readable)
 
     def save_settings(self, path, settings, readable=False):
         """
@@ -776,7 +589,7 @@ class Loadable(Logable):
             raise IOError("File not found {}".format(path))
 
         try:
-            with open(path, "rb") as f:
+            with io.open(path, "r", encoding="utf-8") as f:
                 if path.endswith(".json"):
                     res = self._load_json_file(f)
                 elif path.endswith(".yaml") or path.endswith(".yml"):
@@ -805,7 +618,7 @@ class Loadable(Logable):
             IOError("No path specified to save")
 
         try:
-            with open(path, "wb") as f:
+            with io.open(path, "w", encoding="utf-8") as f:
                 if path.endswith(".json"):
                     self._save_json_file(
                         f,